refactor(svelte): finish CaptureSurface rename; disambiguate plot-px module
Review findings: the mirrored capture-surface test still bound the
component as PlotCaptureSurface (8 stale references); and surface/ held
both pointer.ts and plot-pointer.ts after prefix-dropping erased the old
plot-surface-pointer vs plot-pointer distinction — the coordinate-helper
module is now plot-px.ts, matching what its name always meant.
