+++ packages/svelte/src/lib/surface/surface-state.svelte.ts
@@ -37,7 +37,7 @@ import {
   type IntervalQueryScene,
 } from "../plot-interval-query.js";
 import { BRUSH_SECOND_CORNER_ANNOUNCEMENT } from "../assembly/labels.js";
-import { hitFromCandidate, plotPointFromClient } from "./plot-pointer.js";
+import { hitFromCandidate, plotPointFromClient } from "./plot-px.js";
 import { buildQueuedInspectFrame } from "../plot-surface-inspection-frame.js";
 import {
   resolveSurfaceBlurAction,
